Created interface for table columns cache
