+++ source/Nevermore/ITableColumnsCache.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Nevermore
+{
+    public interface ITableColumnsCache
+    {
+        IEnumerable<string> GetMappingTableColumnNamesSortedWithJsonLast(string schemaName, string tableName);
+    }
+}
